FINAL CHANGES TO MAIL AND SEED DATA
Once you import, be sure to DROP all of you tables. Then run the JPABooks.java and let it run like normal BUT when the main menu pops up, select 3 to exit. Once you exit, refresh your database to get the tables. Once you have your tables, go to the seed-data.sql and run all of the lines in the file. Refresh your database and check that your tables were updated properly. AFTER THAT YOU ARE GOOD TO GOOOOOO